[rename] rename user entity
