+++ .gitignore
@@ -30,3 +30,5 @@ out/
 ### VS Code ###
 .vscode/
 .DS_Store
+
+.yml
