+++ src/main/java/com/nexters/buyornot/domain/model/UserSocialType.java
@@ -1,4 +1,4 @@
-package com.nexters.buyornot.domain.user;
+package com.nexters.buyornot.domain.model;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
